Fix typecheck error on MacOS with python 2.7 when reading plist data
Crossbar fails to start on MacOS running python 2.7 due to a bug in
the way plist data is accessed. The plist method returns an array
with the first value as a dictionary. The current code accesses it
directly as a dictionary, which causes python to throw the type
error. The fix is to simply index it first and then access (this is
already done other places in the code).